requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
drf-nested-routers